Use checked terms as premises for subsequent checks
Retaining checked terms as premises is necessary for RAC to succeed; otherwise
the available information may be insufficient to prove the assertions (e.g.
`rac.Premises.test1`).

In this implementation, every time a term is checked, the term is added to the
set of premises for the following checks, where free variables are substituted
in the term by variables that are bound to the current value of the free
variables, e.g.:

ENV: {x → 0} ASSERT { f x }
ADD PREMISE { let x' = 0 in f x' }

This is not ideal because relations between program states are not represented.
We should instead envisage a VC-gen like behaviour by generating the proof task
during execution. States in the execution should be represented by different
logical variables, and the relations between them are retained (e.g. for
function `f` with postcondition φ(arg, result) the execution in variable
environment {x → x₀} and value environment {x₀ → 0} of `x <- f x` results in
variable environment {x → x₁}, value environment {x₀ → 0, x₁ → 1} and added
premise `φ(x₀, x₁)`. (This approach requires an immutable variable environment,
which isn't currently the case.)

[FEATURE]