TC02-027 Emit attributes RAC:call_id in conversion to Ptree
Normally, result values of function calls are requested during
giant-step RAC by their location. However, lexical, unique locations
can be overwritten using location annotations, which are frequent in
SPARK-generated WhyML programs, resulting in ambiguous locations, and
retrieval of the wrong result values.

As a workaround, this change generates call ids for calls and abstract
expressions, to identify result values in the counterexample uniquely.

In the future, result values should instead be identified by their
structural position in the WhyML AST, to avoid the need for generating
the call ids.

Change-Id: Id4734dc26e0cbdc19289ddeb72f8ec3a42c735f6