TC02-027 Emit attributes RAC:call_id in conversion to Ptree
Normally, result values of function calls are requested during
giant-step RAC by their location. However, location annotations are
frequent in SPARK-generated WhyML programs, with the result that the
location of a call isn't unique anymore. With this change, call ids
are generated for calls and abstract expressions, to identify result
values in the counterexample uniquely.

Change-Id: Id4734dc26e0cbdc19289ddeb72f8ec3a42c735f6