[to upstream] Use attribute RAC:call_id to find return values in giant-step RAC
Normally, result values of function calls are identified during
giant-step RAC by the location of the call. However, location
annotations are frequent generated WhyML programs, with the result
that the locations of calls are not unique anymore.

With this change, the attribute [@RAC:call_id:<id>] can be used on
calls (and exec funs) in the generated programs. Return values are
requested then from the CE during giant-step RAC by the call id <id>
if the call has this annotation, and only otherwise by its location.

Change-Id: Ia145ae5b8fef2a74b7299cf593d074f9fef51287